+++ include/ast/AstUtils.h
@@ -20,8 +20,8 @@ namespace mw {
     std::shared_ptr<Ast> new_string(const std::string &value);
     std::shared_ptr<Ast> new_string(const std::string *value);
 
-    std::shared_ptr<Ast> new_number(int value);
-    std::shared_ptr<Ast> new_number(const int *value);
+    std::shared_ptr<Ast> new_number(int64_t value);
+    std::shared_ptr<Ast> new_number(const int64_t *value);
 
     std::shared_ptr<Ast> new_boolean(bool value);
     std::shared_ptr<Ast> new_boolean(const bool *value);
